// src/types/naming-settings.ts
export const NAMING_SETTINGS_KEY = "namingSettings";

export enum ContentTerms {
  Course = "Course",
  Level = "Level",
  Session = "Session",
  Subjects = "Subjects",
  Modules = "Modules",
  Chapters = "Chapters",
  Slides = "Slides",
  LiveSession = "LiveSession",
}

export enum RoleTerms {
  Admin = "Admin",
  Teacher = "Instructor",
  CourseCreator = "CourseCreator",
  AssessmentCreator = "AssessmentCreator",
  Evaluator = "Evaluator",
  Learner = "Learner",
}
export enum SystemTerms {
  Course = "Courses",
  Level = "Level",
  Session = "Session",
  Subjects = "Subjects",
  Modules = "Modules",
  Chapters = "Chapters",
  Slides = "Slides",
  LiveSession = "Live Session",
  Admin = "Admin",
  Teacher = "Instructor",
  CourseCreator = "Course Creator",
  AssessmentCreator = "Assessment Creator",
  Evaluator = "Evaluator",
  Learner = "Learner",
}